Add a10-3.c and align outputs; use powf everywhere

+++ rtiow/CPP/main14.cpp
@@ -89,7 +89,7 @@ bool refract(const vec3& v, const vec3& n, float ni_over_nt, vec3& refracted) {
 float schlick(float cosine, float ref_idx) {
     float r0 = (1-ref_idx) / (1+ref_idx);
     r0 = r0*r0;
-    return r0 + (1-r0)*pow((1 - cosine),5);
+    return r0 + (1-r0)*powf((1 - cosine),5);
 }
 class dielectric : public material {
     public:
